pyproject.toml
[tool.poetry]
name = "rmqtests"
version = "0.1.0"
description = ""
authors = ["Anton Rodin <freedom.toh3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aio-pika = "^9.5.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
